requirements.txt
flask==3.0.0
flask-cors==4.0.0
livekit==0.13.1
python-dotenv==1.0.0
livekit-agents==1.3.12
